+++ CHANGELOG.md
@@ -29,3 +29,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### [1.0.2] - yusufw2429@gmail.com
 #### Added
 - Graceful shutdown
+- Add health check